complete shift report api

+++ src/main/java/com/retzy/payload/dto/ShiftReportDTO.java
@@ -28,7 +28,7 @@ public class ShiftReportDTO {
 
     private Double netSale;
 
-    private Double totalOrders;
+    private int totalOrders;
 
     private UserDTO cashier;
